Creata la classe Utilities
- Spostato il metodo che effettua un controllo preliminare sulla master password dalla classe Startup alla classe Utilities
- Creati i metodi "openDatabaseFileChooser" e "newDatabaseFileChooser" nella classe Utilities
